fix(cost): fixed get-upload-cost and get-download-cost
