+++ cmd/getallocation.go
@@ -3,36 +3,16 @@ package cmd
 import (
 	"fmt"
 	"log"
-	"math"
 	"os"
-	"time"
 
 	"github.com/0chain/gosdk/core/common"
 	"github.com/0chain/gosdk/zboxcore/blockchain"
-	"github.com/0chain/gosdk/zboxcore/fileref"
 	"github.com/0chain/gosdk/zboxcore/logger"
 	"github.com/0chain/gosdk/zboxcore/sdk"
 	"github.com/0chain/zboxcli/util"
 
 	"github.com/spf13/cobra"
 )
-
-func uploadCostForBlobber(price float64, size int64, data, parity int) (
-	cost common.Balance) {
-
-	var ps = (size + int64(data) - 1) / int64(data)
-	ps = ps * int64(data+parity)
-
-	return common.Balance(price * sizeInGB(ps))
-}
-
-func uploadCostFor1GB(alloc *sdk.Allocation) (cost common.Balance) {
-	for _, d := range alloc.BlobberDetails {
-		cost += uploadCostForBlobber(float64(d.Terms.WritePrice), 1*GB,
-			alloc.DataShards, alloc.ParityShards)
-	}
-	return
-}
 
 // getallocationCmd represents the get allocation info command
 var getallocationCmd = &cobra.Command{
@@ -146,273 +126,12 @@ func sizeInGB(size int64) float64 {
 	return float64(size) / GB
 }
 
-func calculateDownloadCost(alloc *sdk.Allocation, fileSize int64, blocksPerMarker int) common.Balance {
-	chunkSize := fileref.CHUNK_SIZE
-	dataShards := alloc.DataShards
-
-	// singleShardSize collection of data-shards
-	singleShardSize := fileref.CHUNK_SIZE * dataShards
-	totalShards := int(math.Ceil(float64(fileSize) / float64(singleShardSize)))
-
-	// Currently if for example, blocksPerMarker is 10, and there are say 11 shards. First 10 shards will be covered
-	// by first readmarker. Second readmarker is signed with blocksPerMarker number of blocks i.e. 10
-	// so user ends up paying for 20 shards
-	totRMsForEachBlobber := int(math.Ceil(float64(totalShards) / float64(blocksPerMarker)))
-
-	var cost float64
-
-	for _, d := range alloc.BlobberDetails {
-		readPrice, err := d.Terms.ReadPrice.ToToken()
-		if err != nil {
-			log.Fatalf("failed to convert %v to token, %v", d.Terms.ReadPrice, err)
-		}
-		for i := 0; i < totRMsForEachBlobber; i++ {
-			cost += sizeInGB(int64(chunkSize)*int64(blocksPerMarker)) * float64(readPrice)
-		}
-	}
-
-	balance, err := common.ToBalance(cost)
-	if err != nil {
-		log.Fatalf("failed to convert %v to balance, %v", cost, err)
-	}
-	return balance
-}
-func downloadCost(alloc *sdk.Allocation, meta *sdk.ConsolidatedFileMeta, blocksPerMarker int) {
-	if meta.Type != fileref.FILE {
-		log.Fatal("not a file")
-	}
-	// singleShardSize collection of data-shards
-	singleShardSize := fileref.CHUNK_SIZE * alloc.DataShards
-	totalShards := int(math.Ceil(float64(meta.ActualFileSize) / float64(singleShardSize)))
-	requiredBalance := calculateDownloadCost(alloc, meta.ActualFileSize, blocksPerMarker)
-
-	fmt.Printf("%s tokens for %d 64KB blocks (%s) of %s\n", requiredBalance, totalShards*(alloc.DataShards+alloc.ParityShards),
-		common.Size(meta.Size), meta.Path)
-}
-
-// The getDownloadCostCmd returns value in tokens to download a file.
-var getDownloadCostCmd = &cobra.Command{
-	Use:   "get-download-cost",
-	Short: "Get downloading cost",
-	Long:  `Get downloading cost`,
-	Args:  cobra.MinimumNArgs(0),
-	Run: func(cmd *cobra.Command, args []string) {
-
-		var (
-			fflags  = cmd.Flags()
-			allocID string
-			err     error
-		)
-
-		if !fflags.Changed("allocation") {
-			log.Fatal("missing required 'allocation' flag")
-		}
-
-		allocID = cmd.Flag("allocation").Value.String()
-		blocksPerMarker, err := cmd.Flags().GetInt("blocks-per-marker")
-
-		if err != nil {
-			log.Fatal("invalid blocks-per-marker value: ", err)
-		}
-
-		if blocksPerMarker <= 0 {
-			log.Fatal("blocks-per-marker value cannot be <= 0")
-		}
-
-		var (
-			remotePath string
-			authTicket string
-			lookupHash string
-		)
-
-		if fflags.Changed("remotepath") {
-			if remotePath, err = fflags.GetString("remotepath"); err != nil {
-				log.Fatal("invalid 'remotepath' flag: ", err)
-			}
-		}
-
-		if fflags.Changed("authticket") {
-			if authTicket, err = fflags.GetString("authticket"); err != nil {
-				log.Fatal("invalid 'authticket' flag: ", err)
-			}
-		}
-
-		if fflags.Changed("lookuphash") {
-			if lookupHash, err = fflags.GetString("lookuphash"); err != nil {
-				log.Fatal("invalid 'lookuphash' flag: ", err)
-			}
-		}
-
-		if remotePath == "" && authTicket == "" {
-			log.Fatal("'remotepath' or 'authticket' flag required")
-		}
-
-		var (
-			alloc *sdk.Allocation
-			meta  *sdk.ConsolidatedFileMeta
-		)
-
-		if remotePath != "" {
-
-			// by remote path
-
-			if alloc, err = sdk.GetAllocation(allocID); err != nil {
-				log.Fatal("fetching the allocation: ", err)
-			}
-
-			if meta, err = alloc.GetFileMeta(remotePath); err != nil {
-				log.Fatal("can't get file meta: ", err)
-			}
-
-			downloadCost(alloc, meta, blocksPerMarker)
-			return
-		}
-
-		// by authentication ticket
-
-		alloc, err = sdk.GetAllocationFromAuthTicket(authTicket)
-		if err != nil {
-			log.Fatal("can't get allocation object: ", err)
-		}
-		var at = sdk.InitAuthTicket(authTicket)
-
-		if lookupHash == "" {
-			if lookupHash, err = at.GetLookupHash(); err != nil {
-				log.Fatal("can't get lookup hash from auth ticket: ", err)
-			}
-		}
-
-		meta, err = alloc.GetFileMetaFromAuthTicket(authTicket, lookupHash)
-		if err != nil {
-			log.Fatal("can't get file meta: ", err)
-		}
-
-		downloadCost(alloc, meta, blocksPerMarker)
-	},
-}
-
-// The uploadCost for a size and duration (if given). If the duration is zero
-// of less, then it returns upload cost until allocation ends.
-func uploadCost(alloc *sdk.Allocation, size int64, path string,
-	duration time.Duration) {
-
-	var cost common.Balance // total price for size / duration
-
-	for _, d := range alloc.BlobberDetails {
-		cost += uploadCostForBlobber(float64(d.Terms.WritePrice), size,
-			alloc.DataShards, alloc.ParityShards)
-	}
-
-	switch {
-	case duration == 0:
-		fmt.Printf("%s tokens / %s for %s of %s",
-			cost, alloc.TimeUnit, common.Size(size), path)
-	case duration < 0:
-		fmt.Println("allocation expired, 'end' flag can't be used")
-		return
-	default:
-		var dtu = float64(duration) / float64(alloc.TimeUnit)
-		cost = common.Balance(float64(cost) * dtu)
-		fmt.Printf("%s tokens / %s for %s of %s",
-			cost, duration, common.Size(size), path)
-	}
-
-	fmt.Println()
-}
-
-// The getUploadCostCmd returns value in tokens to upload a file.
-var getUploadCostCmd = &cobra.Command{
-	Use:   "get-upload-cost",
-	Short: "Get uploading cost",
-	Long:  `Get uploading cost`,
-	Args:  cobra.MinimumNArgs(0),
-	Run: func(cmd *cobra.Command, args []string) {
-
-		var (
-			fflags   = cmd.Flags()
-			allocID  string
-			duration time.Duration
-			end      bool
-			err      error
-		)
-
-		if !fflags.Changed("allocation") {
-			log.Fatal("missing required 'allocation' flag")
-		}
-
-		allocID = cmd.Flag("allocation").Value.String()
-
-		var localPath string
-
-		if !fflags.Changed("localpath") {
-			log.Fatal("missing requried 'localpath' flag")
-		}
-
-		if localPath, err = fflags.GetString("localpath"); err != nil {
-			log.Fatal("invalid 'localpath' flag: ", err)
-		}
-
-		if localPath == "" {
-			log.Fatal("empty local path")
-		}
-
-		var fi os.FileInfo
-		if fi, err = os.Stat(localPath); err != nil {
-			log.Fatal(err)
-		}
-
-		if !fi.Mode().IsRegular() {
-			log.Fatal("not a regular file")
-		}
-
-		if duration, err = fflags.GetDuration("duration"); err != nil {
-			log.Fatal("invalid 'duration' flag:", err)
-		} else if duration < 0 {
-			log.Fatal("negative duration not allowed: ", duration)
-		}
-
-		if end, err = fflags.GetBool("end"); err != nil {
-			log.Fatal("invalid 'end' flag:", err)
-		}
-
-		var alloc *sdk.Allocation
-		if alloc, err = sdk.GetAllocation(allocID); err != nil {
-			log.Fatal("fetching the allocation: ", err)
-		}
-
-		// until allocation ends
-		if end {
-			var expiry = time.Unix(alloc.Expiration, 0)
-			duration = time.Until(expiry)
-		}
-
-		uploadCost(alloc, fi.Size(), localPath, duration)
-	},
-}
-
 func init() {
 	rootCmd.AddCommand(getallocationCmd)
-	rootCmd.AddCommand(getDownloadCostCmd)
-	rootCmd.AddCommand(getUploadCostCmd)
 
 	getallocationCmd.PersistentFlags().String("allocation", "", "Allocation ID")
 	getallocationCmd.PersistentFlags().Int("blocks-per-marker", 10, "blocks signed per Read Marker")
 	getallocationCmd.MarkFlagRequired("allocation")
 	getallocationCmd.Flags().Bool("json", false, "pass this option to print response as json data")
 
-	dcpf := getDownloadCostCmd.PersistentFlags()
-	dcpf.String("allocation", "", "allocation ID, required")
-	dcpf.String("remotepath", "", "remote path of file")
-	dcpf.Int("blocks-per-marker", 10, "blocks signed per Read Marker")
-	dcpf.String("authticket", "", "authticket")
-	dcpf.String("lookuphash", "", "lookuphash, for the remote file")
-	getDownloadCostCmd.MarkFlagRequired("allocation")
-
-	ucpf := getUploadCostCmd.PersistentFlags()
-	ucpf.String("allocation", "", "allocation ID, required")
-	ucpf.String("localpath", "", "local file path, required")
-	ucpf.Duration("duration", 0, "expected duration keep uploaded file")
-	ucpf.Bool("end", false, "use the duration until allocation ends")
-	getUploadCostCmd.MarkFlagRequired("allocation")
-	getUploadCostCmd.MarkFlagRequired("localpath")
 }
